Adding parse initializer to application
