--- app/src/androidTest/java/com/adil/spott/ParseApplication.java
@@ -1,8 +1,0 @@
-package com.adil.spott;
-
-import android.app.Application;
-
-public class ParseApplication extends Application {
-    public static final String PARSE_APPLICATION_ID = "spott_api";
-    public static final String PARSE_CLIENT_KEY = "fit_360";
-}
